refactor: mypage 리펙토링
